% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accepted_model.R
\name{accepted_model}
\alias{accepted_model}
\title{copy the previous SSC accepted model as the new "base" model}
\usage{
accepted_model(base_year, base_model, year)
}
\arguments{
\item{base_year}{year of the base model}

\item{base_model}{name of the base model (folder name)}

\item{year}{current year}
}
\value{

}
\description{
copy of all base model files, also add a "base_model.txt" file that identifies the model used
}
\examples{
\dontrun{
accepted_model(base_year = 2020, base_model = "m18.2b", year = 2021)
}
}
