% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_error.R
\name{age_error}
\alias{age_error}
\title{Age error analysis}
\usage{
age_error(
  reader_tester = NULL,
  species,
  year,
  admb_home = NULL,
  region = "GOA",
  rec_age = 2,
  plus_age = 45,
  max_age = 100
)
}
\arguments{
\item{species}{= "NORK"}

\item{year}{= year of the assessment}

\item{admb_home}{= location admb exists on your computer - if is "c:/admb" can leave NULL}

\item{region}{= "GOA" (BSAI not currently setup)}

\item{rec_age}{= recruitment age}

\item{plus_age}{= max age for modeling}

\item{max_age}{= max age for age error analysis - default = 100}

\item{read_tester}{= if left blank it looks for the following file:"data/user_input/reader_tester.csv"}
}
\value{

}
\description{
Age error analysis
}
\examples{
ageage(species = "NORK", year = 2020, admb_home = NULL, region = "GOA", rec_age = 2, plus_age = 45, max_age = 100, ...)

}
