% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_f.R
\name{best_f}
\alias{best_f}
\title{best F}
\usage{
best_f(data, m, last_ofl, type = 1, f_ratio = NULL, m2 = NULL)
}
\arguments{
\item{data}{data.frame in form (age, naa, waa, saa)}

\item{m}{natural mortality (for both sexes(\code{type=1}) or for sex 1 (\code{type=2} or \code{type=3} estimates))}

\item{last_ofl}{OFL from the previous assessment}

\item{type}{1 = 1 sex - 1 gear; 2 = 1 sex - 2 gear; 3 = 2 sex - 1 gear - 2m; 4 = 2 sex - 2 gear - 2m}

\item{f_ratio}{is the final ratio of catch between gears in the previous year, as estimated by this year's model - only needed if using multiple gear types}

\item{m2}{is the second estimate of natural mortality}
}
\value{
The estimated F_OFL
}
\description{
Use this year's model to recalculate the F for the previous year that would have produced the specified OFL for the previous year
}
\examples{
\dontrun{
oneone <- data.frame(age = 1:9,
                     naa = c(15.27, 10.85, 0.24, 8.72, 2.49, 10.66, 3.92, 3.71, 2.6),
                     waa = c(0.69, 1.11, 1.54, 1.95, 2.32, 2.64, 2.91, 3.41, 3.33),
                     saa = c(0.01,	0.1	, 0.57, 0.94, 0.99, 1, 1,	1, 1))


onetwo <- data.frame(age = 1:9,
                    naa = c(15.27, 10.85, 0.24, 8.72, 2.49, 10.66, 3.92, 3.71, 2.6),
                    waa = c(0.69, 1.11, 1.54, 1.95, 2.32, 2.64, 2.91, 3.41, 3.33),
                    saa01 = c(0.01,	0.1	, 0.57, 0.94, 0.99, 1, 1,	1, 1),
                    saa02 = c(0.01, 0.06, 0.34, 0.82, 0.97, 1, 1, 1, 1))

twotwo = structure(list(age = 2:10,
                       naa1 = c(7.64, 5.43, 0.12, 4.36, 1.24,5.33, 1.96, 1.85, 1.3),
                       naa2 = c(7.64, 5.42, 0.12, 4.4, 1.27,5.44, 2, 1.89, 1.32),
                       waa1 = c(0.97, 1.46, 1.88, 2.22, 2.48, 2.67, 2.81, 2.91, 2.99),
                       waa2 = c(0.92, 1.48, 2.05, 2.6, 3.09, 3.52, 3.89, 4.19, 4.44),
                       saa11 = c(0.01, 0.1, 0.57, 0.94, 0.99, 1, 1, 1, 1),
                       saa21 = c(0.01, 0.06, 0.34, 0.82, 0.97, 1, 1, 1, 1),
                       saa12 = c(0.12, 0.37, 0.62, 0.82, 0.94, 1, 1, 0.96, 0.89),
                       saa22 = c(0.01, 0.13, 0.38, 0.66, 0.89, 1, 1, 0.92, 0.79)),
                  class = "data.frame", row.names = c(NA, -9L))

type 1
best_f(m = 0.1, last_ofl = 15, data = oneone, type = 1)
type 2
best_f(m = 0.1, last_ofl = 15, data = onetwo, type = 2, f_ratio = 0.1)
type 3
best_f(m = 0.12, last_ofl = 15, data = twotwo, type = 3, f_ratio = 0.1, m2 = 0.08)
type 4
best_f(m = 0.12, last_ofl = 15, data = twotwo, type = 4, f_ratio = 0.1, m2 = 0.08)
}

}
