% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_dat.R
\name{concat_dat}
\alias{concat_dat}
\title{Concatenate a .dat file}
\usage{
concat_dat(
  year,
  species,
  region = "goa",
  model,
  dat_name,
  rec_age,
  plus_age,
  spawn_mo = 5,
  maturity = NULL,
  n_fleets = 1,
  n_ts = NULL,
  n_lls = NULL
)
}
\arguments{
\item{year}{assessment year}

\item{species}{"nork", "rebs", "sabl"}

\item{region}{"goa", "bsai", "everywhere"}

\item{model}{folder that the \code{.tpl} will be in}

\item{dat_name}{what to call the .dat file - ".dat" will be appended to the name}

\item{rec_age}{recruitment age}

\item{plus_age}{plus age group}

\item{spawn_mo}{spawning month}

\item{maturity}{if maturity is from outside the model (should be placed in user_input folder)}

\item{n_fleets}{number of fishing fleets e.g., ll fleet, trawl fleet, default is 1}

\item{n_ts}{number of trawl surveys dafault is 1}

\item{n_lls}{number of longline surveys e.g., domestic and japanes, default is 1}

\item{n_ageage}{number of age error transmission matrices default is 1}

\item{n_sizeage}{number of size at age transmission matrices default is 1}

\item{retro}{not yet implemented}
}
\description{
Concatenate a .dat file
}
\examples{
concat_dat(year = 2020, species = "NORK",  region = "goa", model = "base", dat_name = "goa_nr", rec_age = 2, plus_age = 45)

}
