% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_age_comp.R
\name{fish_age_comp}
\alias{fish_age_comp}
\title{fishery age composition analysis}
\usage{
fish_age_comp(year, fishery = "fsh1", rec_age, plus_age)
}
\arguments{
\item{year}{assessment year}

\item{fishery}{default is fsh1, change if age comps from multiple fisheries (e.g., fsh2)}

\item{rec_age}{recruitment age}

\item{plus_age}{plus age group}
}
\value{

}
\description{
fishery age composition analysis
}
\examples{
\dontrun{
fish_age_comp(year, fishery = "fsh1", rec_age, plus_age)
}
}
