% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_length_comp.R
\name{fish_length_comp}
\alias{fish_length_comp}
\title{Fishery size composition data}
\usage{
fish_length_comp(year, fishery = "fsh1", rec_age, lenbins = NULL)
}
\arguments{
\item{year}{assessment year}

\item{fishery}{default is "fsh1"}

\item{rec_age}{recruitment age}

\item{lenbins}{lenbin file if left NULL it looks for here::here(year, "data", "user_input", "len_bin_labels.csv")}
}
\value{

}
\description{
Fishery size composition data
}
