% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goa_dusk.R
\name{goa_dusk}
\alias{goa_dusk}
\title{raw data query for GOA dusky rockfish}
\usage{
goa_dusk(year, akfin_user, akfin_pwd, afsc_user, afsc_pwd)
}
\arguments{
\item{afsc_pwd}{}
}
\value{

}
\description{
raw data query for GOA dusky rockfish
}
