% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goa_nork.R
\name{goa_nork}
\alias{goa_nork}
\title{raw data query for GOA RE/BS}
\usage{
goa_nork(year, akfin_user, akfin_pwd, afsc_user, afsc_pwd)
}
\arguments{
\item{afsc_pwd}{}
}
\value{

}
\description{
raw data query for GOA RE/BS
}
