% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goa_nork_catch_1961_1992}
\alias{goa_nork_catch_1961_1992}
\title{GOA northern rockfish historical catch data}
\format{
A data frame with 32 observations and 2 variables:
\describe{
\item{year}{year of catch}
\item{catch}{weight of catch in 1,000 t}
}
}
\usage{
goa_nork_catch_1961_1992
}
\description{
A dataset containing trawl gear catch by year for 1961-1992
}
\keyword{datasets}
