% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goa_pop.R
\name{goa_pop}
\alias{goa_pop}
\title{raw data query for GOA POP}
\usage{
goa_pop(year, akfin_user, akfin_pwd, afsc_user, afsc_pwd)
}
\arguments{
\item{year}{assessment year}

\item{akfin_user}{user name}

\item{akfin_pwd}{user password}

\item{afsc_user}{user name}

\item{afsc_pwd}{user password}
}
\value{

}
\description{
raw data query for GOA POP
}
\examples{
\dontrun{
goa_pop(year = 2020, akfin_user, akfin_pwd, afsc_user, afsc_pwd)
}
}
