% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goa_rebs.R
\name{goa_rebs}
\alias{goa_rebs}
\title{raw data query for GOA RE/BS rockfish}
\usage{
goa_rebs(year, akfin_user, akfin_pwd, afsc_user, afsc_pwd)
}
\arguments{
\item{afsc_pwd}{}
}
\value{

}
\description{
raw data query for GOA RE/BS rockfish
}
