% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goa_rebs_catch_1977_2004}
\alias{goa_rebs_catch_1977_2004}
\title{GOA rougheye/blackspotted rockfish historical catch data}
\format{
A data frame with 28 observations and 2 variables:
\describe{
\item{year}{year of catch}
\item{catch}{weight of catch in 1,000 t}
}
}
\usage{
goa_rebs_catch_1977_2004
}
\description{
A dataset containing trawl gear catch by year for 1977-2004
}
\keyword{datasets}
