% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lls_length_comp.R
\name{lls_length_comp}
\alias{lls_length_comp}
\title{Longline length composition analysis}
\usage{
lls_length_comp(YEAR, species = NULL, lenbins = NULL)
}
\arguments{
\item{species}{enter if REBS, otherwise NULL}

\item{lenbins}{lenbin file if left NULL it looks for (year/data/user_input/len_bin_labels.csv")}

\item{year}{assessment year}
}
\value{

}
\description{
Longline length composition analysis
}
