% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lls_rpw.R
\name{lls_rpw}
\alias{lls_rpw}
\title{longline survey relative population weight}
\usage{
lls_rpw(year, area, file = NULL, filter_yrs = NULL)
}
\arguments{
\item{year}{of interest}

\item{area}{"goa", "bsai"}

\item{file}{using an alt-ll survey rpw? (file must have format: year, rpw, cv, sd, lci, uci)}

\item{filter_yrs}{any years that would like removed from survey biomass estimates e.g., 1997 or c(2001, 2003)}
}
\value{

}
\description{
longline survey relative population weight
}
\examples{

}
