% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tsb.R
\name{plot_tsb}
\alias{plot_tsb}
\title{Plot trawl survey biomass (design-based and VAST)}
\usage{
plot_tsb(year, folder, vast = NULL)
}
\arguments{
\item{year}{model year}

\item{folder}{location to place results - a "figs" folder will be created within}

\item{vast}{default is null, otherwise a .csv file placed in the user_input folder}
}
\value{

}
\description{
Plot trawl survey biomass (design-based and VAST)
}
