% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{purrit}
\alias{purrit}
\title{helper function for comp data}
\usage{
purrit(obs, pred = NULL, rec_age, plus_age, comp = "length", lenbins = NULL)
}
\description{
obs  observed data from .rep file
pred predicted data from .rep file (if used)
rec_age  recruitement age
plus_age plus age group
comp age or length - default is length
lenbins set to base unless using alt in which case the file should be in the user_input folder and the name needs to be provided e.g., lengthbins.csv - the column must be named len_bin
}
