% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_fish_length_comp.R
\name{q_fish_length_comp}
\alias{q_fish_length_comp}
\title{fishery length comp data query}
\usage{
q_fish_length_comp(
  year,
  fishery = "fsh1",
  norpac_species,
  area,
  akfin,
  save = TRUE
)
}
\arguments{
\item{year}{assessment year}

\item{fishery}{default is fsh1, change if age comps from multiple fisheries}

\item{norpac_species}{norpac species code}

\item{akfin}{the database to query}

\item{save}{save the file in designated folder}

\item{region}{GOA or BSAI}
}
\value{

}
\description{
fishery length comp data query
}
\examples{

}
