% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_fishery_catch.R
\name{q_fishery_catch}
\alias{q_fishery_catch}
\title{Fishery catch data query}
\usage{
q_fishery_catch(year, fishery = "fsh1", species, area, akfin, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{fishery}{default is fsh1, change if age comps from multiple fisheries}

\item{species}{group species code}

\item{akfin}{the database to query}

\item{save}{save the file in designated folder}

\item{region}{GOA or BSAI}
}
\value{

}
\description{
Fishery catch data query
}
\examples{

}
