% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_fishery_obs.R
\name{q_fishery_obs}
\alias{q_fishery_obs}
\title{Fishery observer data query}
\usage{
q_fishery_obs(year, fishery = "fsh1", norpac_species, area, akfin, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{fishery}{default is fsh1, change if age comps from multiple fisheries}

\item{norpac_species}{norpac species code}

\item{akfin}{the database to query}

\item{save}{save the file in designated folder}

\item{region}{GOA or BSAI}
}
\value{

}
\description{
Fishery observer data query
}
\examples{

}
