% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_lls_biomass.R
\name{q_lls_biomass}
\alias{q_lls_biomass}
\title{Long-line survey biomass data query}
\usage{
q_lls_biomass(year, survey = "goa", afsc_species, afsc, save = TRUE)
}
\arguments{
\item{save}{}
}
\value{

}
\description{
Long-line survey biomass data query
}
