% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_lls_length_comp.R, R/q_ts_length_comp.R
\name{q_lls_length_comp}
\alias{q_lls_length_comp}
\title{Longline survey length comp data query}
\usage{
q_lls_length_comp(year, survey = "goa", afsc_species, area, akfin, save = TRUE)

q_lls_length_comp(year, survey = "goa", afsc_species, area, akfin, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{survey}{default is fsh1, change if age comps from multiple fisheries}

\item{akfin}{the database to query}

\item{save}{save the file in designated folder}

\item{norpac_species}{norpac species code}

\item{region}{GOA or BSAI}
}
\value{



}
\description{
Longline survey length comp data query

trawl survey length comp data query
}
\examples{


}
