% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_ts_age_comp.R
\name{q_ts_age_comp}
\alias{q_ts_age_comp}
\title{trawl survey age comp data query}
\usage{
q_ts_age_comp(year, survey = "goa", afsc_species, area, akfin, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{survey}{default is fsh1, change if age comps from multiple fisheries}

\item{afsc_species}{afsc species code(s)}

\item{akfin}{the database to query}

\item{save}{save the file in designated folder}

\item{region}{GOA or BSAI}
}
\value{

}
\description{
trawl survey age comp data query
}
\examples{

}
