% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_ts_biomass.R
\name{q_ts_biomass}
\alias{q_ts_biomass}
\title{trawl survey biomass data query}
\usage{
q_ts_biomass(year, survey = "goa", afsc_species, afsc, save = TRUE)
}
\arguments{
\item{save}{}
}
\value{

}
\description{
trawl survey biomass data query
}
