% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_ts_length_comp.R
\name{q_ts_length_comp}
\alias{q_ts_length_comp}
\title{trawl survey length comp data query}
\usage{
q_ts_length_comp(year, area = "goa", afsc_species, afsc, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{area}{default is fsh1, change if age comps from multiple fisheries}

\item{save}{save the file in designated folder}

\item{norpac_species}{norpac species code}

\item{akfin}{the database to query}
}
\value{

}
\description{
trawl survey length comp data query
}
\examples{

}
