% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_ts_saa.R
\name{q_ts_saa}
\alias{q_ts_saa}
\title{Trawl survey size at age data query}
\usage{
q_ts_saa(year, area = "goa", afsc_species, afsc, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{area}{default is "goa", change if age comps from other surveys}

\item{afsc_species}{afsc species code}

\item{afsc}{the database to query}

\item{save}{save the file in designated folder}
}
\value{

}
\description{
Trawl survey size at age data query
}
\examples{

}
