% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sabl_catch_1960_1990}
\alias{sabl_catch_1960_1990}
\title{Sablefish historical catch data for 1960-1990}
\format{
A data frame with 62 observations and 3 variables:
\describe{
\item{year}{year of catch}
\item{catch}{weight of catch in 1,000 t}
\item{type}{fishery type is either fixed (includes longline and pot gear) or trawl}
}
}
\usage{
sabl_catch_1960_1990
}
\description{
A dataset containing both fixed gear and trawl gear catch by year
}
\keyword{datasets}
