% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sabl_fixed_abundance}
\alias{sabl_fixed_abundance}
\title{Sablefish historical abundance data}
\format{
A data frame with 121 observations and 5 variables:
\describe{
\item{year}{year}
\item{value}{weight of catch in 1,000 t, or relative abundance}
\item{variable}{rpw, rpn, cpue, or catch}
\item{fleet}{japan or domestic}
\item{gear}{lls = longline survey, llf = longline fishery, tf = trawl fishery}
}
}
\usage{
sabl_fixed_abundance
}
\description{
A dataset containing both fixed gear and trawl gear catch, cpue, rpn & rpw data by year
}
\keyword{datasets}
