% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sabl_fixed_comps}
\alias{sabl_fixed_comps}
\title{Sablefish age and length comp data}
\format{
A data frame with 2010 observations and 7 variables:
\describe{
\item{year}{year}
\item{age}{if relevant, otherwise NA}
\item{comp}{age or length}
\item{fleet}{japan or domestic}
\item{type}{age or length}
\item{gear}{lls = longline survey, tf = trawl fishery, ts = trawl survey}
\item{sex}{male, female or NA}
}
}
\usage{
sabl_fixed_comps
}
\description{
A dataset containing both age and length composition data for multiple fisheries and surveys
}
\keyword{datasets}
