% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sabl_japan_coop_lls}
\alias{sabl_japan_coop_lls}
\title{Sablefish historical CPUE}
\format{
A data frame with 16 observations and 9 variables:
\describe{
\item{year}{year of cpue}
\item{rpn}{relative population number}
\item{rpn_se}{the standard error of rpn}
\item{rpw}{relative population weight}
\item{rpw_se}{the standard error of rpw}
\item{rpn_lci}{lower confidence interval: 2 * se}
\item{rpn_uci}{upper confidence interval: 2 * se}
\item{rpw_lci}{lower confidence interval: 2 * se}
\item{rpw_uci}{upper confidence interval: 2 * se}
}
}
\usage{
sabl_japan_coop_lls
}
\description{
A dataset containing both relative population weight (RPW) and relative population numbers (RPN) from the Japanese cooperative longline fishery for 1979-1994
}
\keyword{datasets}
