% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sabl_japan_llf_rpw}
\alias{sabl_japan_llf_rpw}
\title{Sablefish historical CPUE relative population weight (RPW) from the Japanese longline fishery for 1964-1981}
\format{
A data frame with 18 observations and 5 variables:
\describe{
\item{year}{year of cpue}
\item{rpw}{relative population weight}
\item{se}{the standard error of rpw}
\item{lci}{lower confidence interval: 2 * se}
\item{uci}{upper confidence interval: 2 * se}
}
}
\usage{
sabl_japan_llf_rpw
}
\description{
A dataset containing both cpue data by year
}
\keyword{datasets}
