% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sabl_new_areas}
\alias{sabl_new_areas}
\title{Sablefish defining areas for age samples}
\format{
A data frame with 15 observations and 2 variables:
\describe{
\item{NMFS_AREA}{area number}
\item{AREA_NAME}{assignment for area number}
}
}
\usage{
sabl_new_areas
}
\description{
A dataset containing area definitions
}
\keyword{datasets}
