% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sabl_old_wts_sex}
\alias{sabl_old_wts_sex}
\title{Sablefish fixed weights at length}
\format{
A data frame with 30 observations and 3 variables:
\describe{
\item{length_bin}{length bin groups}
\item{wt_m}{male weight (kg)}
\item{wt_f}{female weight (kg)}
}
}
\usage{
sabl_old_wts_sex
}
\description{
A dataset containing length bins and associated male and female weights
}
\keyword{datasets}
