% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sabl_vessel_lengths}
\alias{sabl_vessel_lengths}
\title{vessel lengths}
\format{
A data frame with 7558 observations and 2 variables:
\describe{
\item{VESSEL_CODE}{vessel id}
\item{LENGTH}{length of the vessel}
}
}
\usage{
sabl_vessel_lengths
}
\description{
Vessel lengths for calculating whale depredation rates in the sablefish fishery
}
\keyword{datasets}
