% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sabl_waa}
\alias{sabl_waa}
\title{Sablefish fixed weights at age}
\format{
A data frame with 120 observations and 3 variables:
\describe{
\item{waa}{length bin groups}
\item{wt_m}{male weight (kg)}
\item{wt_f}{female weight (kg)}
}
}
\usage{
sabl_waa
}
\description{
A dataset containing weights at age for two time blocks, by sex
}
\keyword{datasets}
