% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sablefish.R
\name{sablefish}
\alias{sablefish}
\title{raw data query for sablefish}
\usage{
sablefish(year, akfin_user, akfin_pwd, afsc_user, afsc_pwd)
}
\arguments{
\item{afsc_pwd}{}
}
\value{

}
\description{
raw data query for sablefish
}
\examples{
\donttest{
sablefish(year, akfin_user, akfin_pwd, afsc_user, afsc_pwd)
}
}
