% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup folder structure}
\usage{
setup(year)
}
\arguments{
\item{year}{}
}
\value{

}
\description{
Creates a common folder structure for ADMB assessments
}
\examples{
\dontrun{
setup(2020)
}
}
