% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_proj.R
\name{setup_proj}
\alias{setup_proj}
\title{Title}
\usage{
setup_proj(year, model, species, area = "goa", off_yr = NULL)
}
\arguments{
\item{year}{assessment year}

\item{model}{the folder the ADMB model is in}

\item{species}{identifier for file names}

\item{area}{e.g. "goa"}

\item{off_yr}{is this a partial year (TRUE) otherwise NULL}
}
\value{

}
\description{
Title
}
