% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_at_age.R
\name{size_at_age}
\alias{size_at_age}
\title{size at age analysis}
\usage{
size_at_age(year, admb_home = NULL, rec_age, lenbins = NULL)
}
\arguments{
\item{year}{analysis year}

\item{admb_home}{location admb exists on your computer}

\item{rec_age}{recruitment age}

\item{lenbins}{length bin file}

\item{max_age}{max age for age error analysis - default = 100}
}
\value{

}
\description{
size at age analysis
}
