% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_age_comp.R
\name{ts_age_comp}
\alias{ts_age_comp}
\title{trawl survey age comp analysis}
\usage{
ts_age_comp(year, area = "goa", rec_age, plus_age, rmv_yrs = NULL)
}
\arguments{
\item{year}{assessment year}

\item{area}{default is "goa"}

\item{rec_age}{recruitment age}

\item{plus_age}{plus group age}

\item{rmv_yrs}{any survey years to exclude}
}
\value{

}
\description{
trawl survey age comp analysis
}
\examples{
ts_age_comp(year = 2020, rec_age = 2, plus_age = 45)
}
