% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_biomass.R
\name{ts_biomass}
\alias{ts_biomass}
\title{survey biomass}
\usage{
ts_biomass(year, area = "goa", file = NULL)
}
\arguments{
\item{year}{of interest}

\item{area}{e.g., "goa"}

\item{file}{if not using the design-based abundance, the file name must be stated (e.g. "GAP_VAST.csv")}
}
\value{

}
\description{
survey biomass
}
\examples{

}
