% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_length_comp.R
\name{ts_length_comp}
\alias{ts_length_comp}
\title{trawl survey length composition analysis}
\usage{
ts_length_comp(year, survey = "goa", lenbins = NULL)
}
\arguments{
\item{year}{assessment year}

\item{survey}{survey area default = "goa"}

\item{lenbins}{lenbin file if left NULL it looks for (year/data/user_input/len_bins.csv")}
}
\value{

}
\description{
trawl survey length composition analysis
}
\examples{

}
