% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_at_age.R
\name{weight_at_age}
\alias{weight_at_age}
\title{estimate allometric relationship and weight-at-age}
\usage{
weight_at_age(year, admb_home = NULL, rec_age, area = "goa")
}
\arguments{
\item{year}{model year}

\item{admb_home}{= location admb exists on your computer}

\item{rec_age}{recruitment age}

\item{area}{currently fixed at "goa"}
}
\value{

}
\description{
estimate allometric relationship and weight-at-age
}
\examples{
weight_at_age(year = 2020, admb_home = "C:/Program Files (x86)/ADMB-12.1", rec_age = 2)
}
