% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_catch.R
\name{clean_catch}
\alias{clean_catch}
\title{clean up catch data}
\usage{
clean_catch(
  year,
  fishery = "fsh1",
  TAC = c(3333, 2222, 1111),
  fixed_catch = NULL
)
}
\arguments{
\item{year}{\itemize{
\item year of assessment
}}

\item{fishery}{identify the fishery default is "fsh1"}

\item{TAC}{\itemize{
\item last three TAC in form: c(year-3, year-2, year-1)
}}

\item{fixed_catch}{\itemize{
\item if catch is frozen place the file in user_input folder (format: Year, Catch)
}}
}
\value{

}
\description{
clean up catch data
}
\examples{
clean_catch(year, TAC = c(2874, 2756, 3100), fixed_catch = "catch_1961-1992.csv")

}
