% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goa_dusk.R
\name{goa_dusk}
\alias{goa_dusk}
\title{raw data query for GOA dusky rockfish}
\usage{
goa_dusk(year, akfin_user, akfin_pwd, afsc_user, afsc_pwd, off_yr = NULL)
}
\arguments{
\item{year}{assessment year}

\item{akfin_user}{user name}

\item{akfin_pwd}{user password}

\item{afsc_user}{user name}

\item{afsc_pwd}{user password}

\item{off_yr}{if this is an off-year assessment change to TRUE}
}
\value{

}
\description{
raw data query for GOA dusky rockfish
}
\examples{
\dontrun{
goa_disk(year, akfin_user, akfin_pwd, afsc_user, afsc_pwd)
}
}
