% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goa_nork.R
\name{goa_nork}
\alias{goa_nork}
\title{raw data query for GOA RE/BS}
\usage{
goa_nork(year, akfin_user, akfin_pwd, afsc_user, afsc_pwd, off_yr = NULL)
}
\arguments{
\item{year}{assessment year}

\item{akfin_user}{user name}

\item{akfin_pwd}{user password}

\item{afsc_user}{user name}

\item{afsc_pwd}{user password}

\item{off_yr}{if this is an off-year assessment change to TRUE}
}
\value{

}
\description{
raw data query for GOA RE/BS
}
\examples{
\dontrun{
goa_nork(year = 2020, akfin_user, akfin_pwd, afsc_user, afsc_pwd)
}
}
