% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lls_rpn.R
\name{lls_rpn}
\alias{lls_rpn}
\title{longline survey relative population numbers}
\usage{
lls_rpn(year, survey, file = NULL, filter_yrs = NULL)
}
\arguments{
\item{year}{of interest}

\item{survey}{"goa", "bsai"}

\item{file}{using an alt-ll survey rnp? (file must have format: year, rnp, cv, sd, lci, uci)}

\item{filter_yrs}{any years that would like removed from survey biomass estimates e.g., 1997 or c(2001, 2003)}
}
\value{

}
\description{
longline survey relative population numbers
}
\examples{

}
