% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_lls_biomass.R
\name{q_lls_biomass}
\alias{q_lls_biomass}
\title{longline survey biomass data query}
\usage{
q_lls_biomass(year, area = "goa", afsc_species, akfin, save = TRUE)
}
\arguments{
\item{save}{}
}
\value{

}
\description{
longline survey biomass data query
}
