% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_lls_length_comp.R
\name{q_lls_length_comp}
\alias{q_lls_length_comp}
\title{Longline survey length comp data query}
\usage{
q_lls_length_comp(year, area = "goa", afsc_species, akfin, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{area}{default is fsh1, change if age comps from multiple fisheries}

\item{afsc_species}{afsc species code}

\item{akfin}{the database to query}

\item{save}{save the file in designated folder}
}
\value{

}
\description{
Longline survey length comp data query
}
\examples{

}
