% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_ts_age_comp.R
\name{q_ts_age_comp}
\alias{q_ts_age_comp}
\title{trawl survey age comp data query}
\usage{
q_ts_age_comp(year, area = "goa", afsc_species, afsc, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{area}{default is goa, change if age comps from multiple surveys}

\item{afsc_species}{afsc species code(s)}

\item{afsc}{the database to query}

\item{save}{save the file in designated folder}
}
\value{

}
\description{
trawl survey age comp data query
}
\examples{

}
