% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_at_age.R
\name{size_at_age}
\alias{size_at_age}
\title{compute size at age (length at age and length sd)}
\usage{
size_at_age(year, admb_home, rec_age, lenbins = NULL)
}
\arguments{
\item{year}{analysis year}

\item{admb_home}{location admb exists on your computer}

\item{rec_age}{recruitment age}

\item{lenbins}{length bin file}

\item{max_age}{max age for age error analysis - default = 100}
}
\value{

}
\description{
compute size at age (length at age and length sd)
}
